Extract session autosave coordinator
