+++ cmuxTests/SessionPersistenceTests.swift
@@ -856,103 +856,6 @@ final class SessionPersistenceTests: XCTestCase {
         )
     }
 
-    @MainActor
-    func testSessionAutosaveCoordinatorSerializesTicksAndIgnoresStaleFinish() throws {
-        let coordinator = SessionAutosaveCoordinator()
-        var startedCount = 0
-        var firstRunToken: SessionAutosaveCoordinator.RunToken?
-
-        XCTAssertTrue(
-            coordinator.beginTick { runToken in
-                startedCount += 1
-                firstRunToken = runToken
-                return Task {}
-            }
-        )
-        XCTAssertFalse(
-            coordinator.beginTick { _ in
-                startedCount += 1
-                return Task {}
-            }
-        )
-
-        coordinator.finishTick(runToken: SessionAutosaveCoordinator.RunToken())
-        XCTAssertFalse(
-            coordinator.beginTick { _ in
-                startedCount += 1
-                return Task {}
-            }
-        )
-
-        coordinator.finishTick(runToken: try XCTUnwrap(firstRunToken))
-        XCTAssertTrue(
-            coordinator.beginTick { _ in
-                startedCount += 1
-                return Task {}
-            }
-        )
-        coordinator.cancelInFlightTick()
-
-        XCTAssertEqual(startedCount, 2)
-    }
-
-    @MainActor
-    func testSessionAutosaveCoordinatorTracksTypingQuietPeriod() throws {
-        let coordinator = SessionAutosaveCoordinator()
-
-        XCTAssertNil(coordinator.remainingTypingQuietPeriod(quietPeriod: 0.65, nowUptime: 100))
-        coordinator.recordTypingActivity(nowUptime: 100)
-
-        XCTAssertEqual(
-            try XCTUnwrap(coordinator.remainingTypingQuietPeriod(quietPeriod: 0.65, nowUptime: 100.25)),
-            0.40,
-            accuracy: 0.001
-        )
-        XCTAssertNil(coordinator.remainingTypingQuietPeriod(quietPeriod: 0.65, nowUptime: 101))
-    }
-
-    @MainActor
-    func testSessionAutosaveCoordinatorTracksPersistedFingerprint() {
-        let coordinator = SessionAutosaveCoordinator()
-        let now = Date()
-
-        XCTAssertFalse(
-            coordinator.shouldSkipSaveForUnchangedFingerprint(
-                isTerminatingApp: false,
-                includeScrollback: false,
-                currentFingerprint: 42,
-                now: now,
-                maximumAutosaveSkippableInterval: 60
-            )
-        )
-
-        coordinator.recordSuccessfulSave(
-            isTerminatingApp: false,
-            includeScrollback: false,
-            persistedAt: now,
-            fingerprint: 42
-        )
-
-        XCTAssertTrue(
-            coordinator.shouldSkipSaveForUnchangedFingerprint(
-                isTerminatingApp: false,
-                includeScrollback: false,
-                currentFingerprint: 42,
-                now: now.addingTimeInterval(10),
-                maximumAutosaveSkippableInterval: 60
-            )
-        )
-        XCTAssertFalse(
-            coordinator.shouldSkipSaveForUnchangedFingerprint(
-                isTerminatingApp: false,
-                includeScrollback: false,
-                currentFingerprint: 43,
-                now: now.addingTimeInterval(10),
-                maximumAutosaveSkippableInterval: 60
-            )
-        )
-    }
-
     func testSessionAutosaveFingerprintIncludesRestorableAgentMetadata() throws {
         let workspaceId = UUID()
         let panelId = UUID()
